refactor(labor): extract shared localWindow helper
Move the restaurant-local window math out of useSplhData into
src/lib/localDateWindow.ts so useLaborSalesAnalytics can reuse the exact
same window. No behavior change.
