+++ src/hooks/useSplhData.ts
@@ -3,6 +3,7 @@ import { fromZonedTime } from 'date-fns-tz';
 
 import { supabase } from '@/integrations/supabase/client';
 
+import { localWindow } from '@/lib/localDateWindow';
 import type { SplhSaleRow } from '@/lib/splhAnalytics';
 import type { TimePunch } from '@/types/timeTracking';
 
@@ -81,27 +82,6 @@ async function fetchAllPunches(
     if (!data || data.length < PAGE) return { rows, capped: false };
   }
   return { rows, capped: true };
-}
-
-/**
- * Window boundaries derived from the restaurant-local "today", not host/UTC
- * `new Date()` (§5 S-min1). `endStr` is today's date in `tz`; `startStr` is
- * `weeks` weeks earlier. Dates are formatted as plain YYYY-MM-DD (no time
- * component), matching `unified_sales.sale_date`'s column type.
- */
-function localWindow(tz: string, weeks: number): { startStr: string; endStr: string } {
-  const now = new Date();
-  const fmt = new Intl.DateTimeFormat('en-CA', {
-    timeZone: tz,
-    year: 'numeric',
-    month: '2-digit',
-    day: '2-digit',
-  });
-  const endStr = fmt.format(now); // YYYY-MM-DD in tz (en-CA locale formats this way)
-  const [y, m, d] = endStr.split('-').map(Number);
-  const start = new Date(Date.UTC(y, m - 1, d));
-  start.setUTCDate(start.getUTCDate() - weeks * 7);
-  return { startStr: start.toISOString().slice(0, 10), endStr };
 }
 
 /**
